refactor(java): move JSON subtype codecs
